infra/pyproject.toml
[project]
name = "novascan-infra"
version = "0.1.0"
description = "NovaScan CDK infrastructure"
requires-python = ">=3.13"
dependencies = [
    "aws-cdk-lib>=2.170.0",
    "constructs>=10.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
]

[tool.hatch.build.targets.wheel]
packages = ["stacks", "cdkconstructs"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
